+++ src/lib/network_scanner/__init__.py
@@ -1,0 +1,3 @@
+#! /usr/bin/python
+
+pass
