+++ src/lib/ssl_scanner/__init__.py
@@ -1,0 +1,3 @@
+#! /usr/bin/python
+
+pass
